computer_interaction_tools/cli_oi_tool/pyproject.toml
[tool.poetry]
name = "mcp_open_interpreter"
version = "0.1.0"
description = "MCP container exposing Open Interpreter via FastMCP as CLI tool"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.12.10"
fastmcp = "^2.0.0"
open-interpreter = "0.4.3"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"